+++ src/hamocc/common/mo_aggregates.f90
@@ -150,6 +150,11 @@ MODULE mo_aggregates
 
 !     REAL(wp), INTENT(in)           :: lon(bgc_nproma), lat(bgc_nproma)
 
+#ifdef __LVECTOR__
+     INTEGER :: max_klevs
+     max_klevs = MAXVAL(klev(start_idx:end_idx))
+#endif
+
      ! molecular dynamic viscosity
      CALL calc_dynvis(aggr_mem%dynvis, klev, start_idx, end_idx, pddpo, ppo, ptho, psao)
 
@@ -158,10 +163,18 @@ MODULE mo_aggregates
      ! calculate the mean sinking velocity of aggregates
      CALL ws_Re_approx(aggr_mem, klev, start_idx, end_idx, pddpo)
 
+#ifdef __LVECTOR__
+     !NEC$ nomove
+     DO k = 1, max_klevs
+        !NEC$ nomove
+        DO j = start_idx, end_idx
+            IF( (pddpo(j,k) > EPSILON(0.5_wp)) .and. (k <= klev(j)) ) THEN
+#else
      DO j = start_idx, end_idx
         kpke=klev(j)
         DO k = 1,kpke
            IF(pddpo(j,k) > EPSILON(0.5_wp))THEN
+#endif
 
 !              aggdiag(j,k,kwsagg)    = ws_agg(j,k) * 86400._wp / dtbgc ! conversion  m/time_step   to  m/d for output
 !              aggdiag(j,k,kdynvis)   = dynvis(j,k)
@@ -231,6 +244,11 @@ MODULE mo_aggregates
                        &  av_rhof_V(:,:)
      REAL(wp), POINTER :: aggdiag(:,:,:)    ! 3d concentration EU
 
+#ifdef __LVECTOR__
+     INTEGER :: max_klevs
+     max_klevs = MAXVAL(klev(start_idx:end_idx))
+#endif
+
        av_dp              =>  aggr_mem%av_dp
        av_rho_p           =>  aggr_mem%av_rho_p
        df_agg             =>  aggr_mem%df_agg
@@ -243,6 +261,13 @@ MODULE mo_aggregates
        av_rhof_V          =>  aggr_mem%av_rhof_V
        aggdiag            =>  aggr_mem%aggdiag
 
+#ifdef __LVECTOR__
+     !NEC$ nomove
+     DO k = 1, max_klevs
+        !NEC$ nomove
+        DO j = start_idx, end_idx
+            IF( (pddpo(j,k) > EPSILON(0.5_wp)) .and. (k <= klev(j)) ) THEN
+#else
      DO j = start_idx, end_idx
         kpke=klev(j)
         IF(kpke > 0)THEN
@@ -250,6 +275,7 @@ MODULE mo_aggregates
         DO k = 1,kpke
 
          IF(pddpo(j,k) > EPSILON(0.5_wp))THEN
+#endif
 
               n_det   = 0._wp ! number of primary particles
               n_opal  = 0._wp
@@ -412,7 +438,9 @@ MODULE mo_aggregates
 
             ENDIF
          ENDDO
+#ifndef __LVECTOR__
          ENDIF
+#endif
       ENDDO
 
 
@@ -421,7 +449,13 @@ MODULE mo_aggregates
 !   CALL max_agg_diam(kpie, kpje, kpke, pddpo)
    CALL max_agg_diam(aggr_mem, klev, start_idx, end_idx, pddpo)
 
-
+#ifdef __LVECTOR__
+   !NEC$ nomove
+   DO k = 1, max_klevs
+      !NEC$ nomove
+      DO j = start_idx, end_idx
+          IF( (pddpo(j,k) > EPSILON(0.5_wp)) .and. (k <= klev(j)) ) THEN
+#else
      DO j = start_idx, end_idx
         kpke=klev(j)
         IF(kpke > 0)THEN
@@ -429,6 +463,7 @@ MODULE mo_aggregates
         DO k = 1,kpke
 
          IF(pddpo(j,k) > EPSILON(0.5_wp))THEN
+#endif
 
              ! mass factor  ! mm: only used to calculate number of aggregates n_agg; see unmodularized MAGO version
 !             mf = mass_factor(av_dp(j,k), df_agg(j,k), av_rho_p(j,k))
@@ -473,7 +508,9 @@ MODULE mo_aggregates
 
             ENDIF
          ENDDO
+#ifndef __LVECTOR__
          ENDIF
+#endif
       ENDDO
 
 
@@ -501,17 +538,29 @@ MODULE mo_aggregates
      INTEGER, INTENT(in)            :: end_idx                !< end index  for j loop  (ICON cells, MPIOM lat dir)
      REAL(wp), INTENT(in), TARGET   :: pddpo(bgc_nproma,bgc_zlevs)      !< size of scalar grid cell (3rd dimension) [m]
 
+#ifdef __LVECTOR__
+     INTEGER :: max_klevs
+     max_klevs = MAXVAL(klev(start_idx:end_idx))
 
+     !NEC$ nomove
+     DO k = 1, max_klevs
+        !NEC$ nomove
+        DO j = start_idx, end_idx
+            IF( (pddpo(j,k) > EPSILON(0.5_wp)) .and. (k <= klev(j)) ) THEN
+#else
      DO j = start_idx, end_idx
         kpke=klev(j)
         IF(kpke > 0)THEN
         DO k = 1,kpke
            IF(pddpo(j,k) > EPSILON(0.5_wp))THEN
+#endif
               ! ws_Re is a function
               aggr_mem%ws_agg(j,k) = ws_Re(aggr_mem,j,k)
            ENDIF
         ENDDO
-        ENDIF
+#ifndef __LVECTOR__
+        ENDIF ! kpke > 0
+#endif
      ENDDO
 
 
@@ -582,6 +631,7 @@ MODULE mo_aggregates
 
     ! Local
     REAL(wp) :: d_Re01, d_Re10, d_low, ws_agg_ints
+#ifndef __LVECTOR__
     REAL(wp),POINTER :: av_dp(:,:),               &  ! mean primary particle diameter
                        &  av_rho_p(:,:),           &  ! mean primary particle density
                        &  df_agg(:,:),             &  ! fractal dimension of aggregates
@@ -652,6 +702,62 @@ MODULE mo_aggregates
             & - av_dp(j,k)**(1._wp + df_agg(j,k) - b_agg(j,k)))  &
             & / (1._wp + df_agg(j,k) - b_agg(j,k))))*dtbgc   ! (m/s -> m/d)  *dtb
 
+#else
+    ! ******************************************************************************************************************
+    ! This code block, which accesses structure members directly, is used to bypass the NEC compiler
+    ! limitation: "*** Structure assignment obstructs vectorization."
+    ! ******************************************************************************************************************
+
+    d_Re01 = get_dRe(aggr_mem%dynvis(j,k), aggr_mem%df_agg(j,k), aggr_mem%av_rho_p(j,k), aggr_mem%av_dp(j,k), &
+      & AJ1, BJ1, 0.1_wp)
+
+    ! Re=10
+    d_Re10 = get_dRe(aggr_mem%dynvis(j,k), aggr_mem%df_agg(j,k), aggr_mem%av_rho_p(j,k), aggr_mem%av_dp(j,k), &
+      & AJ2, BJ2, 10._wp)
+    d_low = aggr_mem%av_dp(j,k)
+
+    ws_agg_ints = 0._wp
+    IF(aggr_mem%Lmax_agg(j,k) >= d_Re01)THEN ! Re > 0.1
+                                    ! - collect full range up to
+                                    ! 0.1, (dp->d_Re1) and set lower bound to
+                                    ! Re=0.1 val
+                                    ! aj=AJ1, bj=1
+      ws_agg_ints = get_ws_agg_integral(aggr_mem%dynvis(j,k), aggr_mem%av_rho_p(j,k), aggr_mem%av_dp(j,k), &
+        & aggr_mem%df_agg(j,k), aggr_mem%b_agg(j,k), &
+        & AJ1, BJ1, aggr_mem%av_dp(j,k), d_Re01)
+        d_low = d_Re01
+    ENDIF
+
+    IF(aggr_mem%Lmax_agg(j,k) >= d_Re10)THEN ! Re > 10
+                                         ! - collect full range Re=0.1-10 (d_Re1-> d_Re2)
+                                         ! and set lower bound to
+                                         ! Re=10 val
+                                         ! aj=AJ2, bj=0.871
+        ws_agg_ints = ws_agg_ints  + &
+         get_ws_agg_integral(aggr_mem%dynvis(j,k), aggr_mem%av_rho_p(j,k), aggr_mem%av_dp(j,k), &
+          & aggr_mem%df_agg(j,k), aggr_mem%b_agg(j,k), &
+          AJ2, BJ2, d_Re01, d_Re10)
+        d_low = d_Re10
+    ENDIF
+
+    IF(d_low < d_Re01)THEN ! Re<0.1 and Lmax < d_Re1
+        ws_agg_ints = get_ws_agg_integral(aggr_mem%dynvis(j,k), aggr_mem%av_rho_p(j,k), aggr_mem%av_dp(j,k), &
+          & aggr_mem%df_agg(j,k), aggr_mem%b_agg(j,k), &
+          & AJ1, BJ1, aggr_mem%av_dp(j,k), aggr_mem%Lmax_agg(j,k))
+    ELSE ! Re > 10, aj=AJ3, bj=BJ3
+        ws_agg_ints = ws_agg_ints + &
+          & get_ws_agg_integral(aggr_mem%dynvis(j,k), aggr_mem%av_rho_p(j,k), aggr_mem%av_dp(j,k), &
+          & aggr_mem%df_agg(j,k), aggr_mem%b_agg(j,k), AJ3, BJ3, d_low, aggr_mem%Lmax_agg(j,k))
+    ENDIF
+
+    ! concentration-weighted mean sinking velocity
+    ws_Re = (ws_agg_ints &
+            & /((aggr_mem%Lmax_agg(j,k)**(1._wp + aggr_mem%df_agg(j,k) - aggr_mem%b_agg(j,k))  &
+            & - aggr_mem%av_dp(j,k)**(1._wp + aggr_mem%df_agg(j,k) - aggr_mem%b_agg(j,k)))  &
+            & / (1._wp + aggr_mem%df_agg(j,k) - aggr_mem%b_agg(j,k))))*dtbgc   ! (m/s -> m/d)  *dtb
+
+#endif
+
   END FUNCTION ws_Re
 
   !=====================================================================================
@@ -676,7 +782,16 @@ MODULE mo_aggregates
      ! Local variables
      REAL(wp) :: nu_vis
 
+#ifdef __LVECTOR__
+     INTEGER :: max_klevs
+     max_klevs = MAXVAL(klev(start_idx:end_idx))
 
+     !NEC$ nomove
+     DO k = 1, max_klevs
+        !NEC$ nomove
+        DO j = start_idx, end_idx
+            IF( (pddpo(j,k) > EPSILON(0.5_wp)) .and. (k <= klev(j)) ) THEN
+#else
      DO j = start_idx, end_idx
         kpke=klev(j)
         IF(kpke > 0)THEN
@@ -684,6 +799,7 @@ MODULE mo_aggregates
         DO k = 1,kpke
 
            IF(pddpo(j,k) > EPSILON(0.5_wp))THEN
+#endif
 
 ! ori:
 !              Lmax_agg(j,k) = ((agg_Re_crit * 18._wp *  dynvis(j,k) * dynvis(j,k) / rhoref_water)&
@@ -699,7 +815,9 @@ MODULE mo_aggregates
 
            ENDIF
         ENDDO
+#ifndef __LVECTOR__
         ENDIF
+#endif
      ENDDO
 
 
@@ -772,7 +890,16 @@ MODULE mo_aggregates
      ! Local variables
      REAL(wp):: press_val  ! Pascal/rho -> dbar
 
+#ifdef __LVECTOR__
+     INTEGER :: max_klevs
+     max_klevs = MAXVAL(klev(start_idx:end_idx))
 
+     !NEC$ nomove
+     DO k = 1, max_klevs
+        !NEC$ nomove
+        DO j = start_idx, end_idx
+            IF( (pddpo(j,k) > EPSILON(0.5_wp)) .and. (k <= klev(j)) ) THEN
+#else
      DO j = start_idx, end_idx
         kpke=klev(j)
         IF(kpke > 0)THEN
@@ -780,6 +907,7 @@ MODULE mo_aggregates
         DO k = 1,kpke
 
            IF(pddpo(j,k) > EPSILON(0.5_wp))THEN
+#endif
 
 !              press_val = ppo(j,k) * rhoref_water * 1.e-5_wp ! Pascal/rho -> dbar
               press_val = ppo(j,k) * rhoref_water/g * 1.e-4_wp ! mm
@@ -801,7 +929,9 @@ MODULE mo_aggregates
 
            ENDIF
         ENDDO
+#ifndef __LVECTOR__
         ENDIF
+#endif
      ENDDO