+++ src/hamocc/icon_specific/mo_bgc_icon_comm.f90
@@ -106,7 +106,7 @@
       REAL(wp),INTENT(inout) :: pco2flx(nproma)
       LOGICAL, INTENT(IN), OPTIONAL :: lacc
 
-      INTEGER :: jc, jk, kpke
+      INTEGER :: jc, jk, kpke, max_klevs
       INTEGER :: start_idx, end_idx
       INTEGER :: itrac
       LOGICAL :: lzacc
@@ -116,6 +116,8 @@
      ! CALL message(TRIM(routine), 'start' )
 
      CALL set_acc_host_or_device(lzacc, lacc)
+
+#ifndef __LVECTOR__
 
       !$ACC PARALLEL DEFAULT(PRESENT) ASYNC(1) IF(lzacc)
       !$ACC LOOP GANG VECTOR
@@ -133,6 +135,28 @@
         ENDIF
       ENDDO
       !$ACC END PARALLEL
+
+#else
+
+    max_klevs = MAXVAL(klevs(start_idx:end_idx))
+
+    DO jc = start_idx , end_idx
+        IF (pddpo(jc, 1) .GT. EPSILON(0.5_wp)) THEN
+            pco2flx(jc) = local_bgc_mem%bgcflux(jc,kcflux_cpl) * molw_co2
+        END IF
+    END DO
+
+    DO itrac = 1, n_bgctra
+        DO jk = 1, max_klevs
+            DO jc = start_idx , end_idx
+                IF ( (jk <= klevs(jc)) .AND. (pddpo(jc, 1) .GT. EPSILON(0.5_wp)) ) THEN
+                    ptracer(jc,jk,jb,itrac) = local_bgc_mem%bgctra(jc,jk,itrac)
+                END IF
+            END DO
+        END DO
+    END DO ! itrac
+
+#endif
 
       END SUBROUTINE
 
@@ -245,17 +269,31 @@
         END DO
       END DO
 
+#ifndef __LVECTOR__
       !$ACC LOOP GANG VECTOR COLLAPSE(2)
       DO jk =1,max_klevs
         DO jc=start_index,end_index
           IF (pddpo(jc, 1) .GT. EPSILON(0.5_wp) .and. jk <= klevs(jc)) THEN
             !$ACC LOOP SEQ
             DO itrac = 1,n_bgctra
+#else
+    DO itrac = 1, n_bgctra
+        DO jk = 1, max_klevs
+            DO jc = start_index, end_index
+                IF (pddpo(jc, 1) .GT. EPSILON(0.5_wp) .and. jk <= klevs(jc)) THEN
+#endif
               local_bgc_mem%bgctra(jc,jk,itrac)=ptracer(jc,jk,jb,itrac)
+#ifndef __LVECTOR__
             END DO
           END IF
         END DO
       END DO
+#else
+                END IF
+            END DO
+        END DO
+    END DO
+#endif
 
       IF (lsediment_only) THEN
         !$ACC LOOP GANG VECTOR